requirements.txt
# API & config
fastapi>=0.115.0,<0.136.0
uvicorn[standard]>=0.30.0,<0.43.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.5.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.9

# Google Gemini (Gen AI SDK)
google-genai>=1.0.0

# Solana / Anchor: anchorpy 0.21+ использует solders (Transaction и т.д.), совместим с solana-py 0.36+
solana>=0.36.0
solders>=0.21.0
anchorpy>=0.21.0

# MCP (stdio + HTTP/SSE)
mcp>=1.0.0
sse-starlette>=2.0.0,<3.0.0

# Tests
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
